requirements.txt
Django==5.1.7
djangorestframework==3.14.0
django-unfold==0.65.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-decouple==3.8
drf-yasg==1.21.5
django-cors-headers==4.3.1
django-filter==23.3
dotenv
Pillow